requirements.txt
openai>=1.40.0
google-generativeai>=0.7.2
pyyaml>=6.0.1
pypdf>=4.2.0
pymupdf>=1.24.0
python-pptx>=1.0.0
python-docx>=1.1.2
openpyxl>=3.1.5
python-dotenv>=1.0.1
watchdog>=4.0.1
pywin32>=306; platform_system=="Windows"
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
keyring>=25.0.0
pywebview>=5.0
sqlite-vec>=0.1.1
